pyproject.toml
[tool.poetry]
name = "bitwarden"
version = "0.1.0"
description = ""
authors = ["J-Lehrack <jjlehrack899@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
requests = "2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
